proposal: Move elektraLookupOptions into kdbhelper
- Having them in kdbprivate causes a declaration
conflict for elektraOpen and elektraClose
in the intercept binding.
